refact: refactor codebase to improve code quality

+++ src/scripts/terminator/terminator.ts
@@ -1,5 +1,5 @@
 import * as db from '../util/database'
-import { hideElements, isElementHidden, once, showElements } from "../util"
+import { /* hideElements, isElementHidden, */ greyInElements, greyOutElements, hideElements, once, /* showElements */ } from "../util"
 import { getI18nMessage as _ } from '../util/i18n'
 
 export abstract class Terminator {
@@ -38,7 +38,8 @@ export abstract class ListedTerminator extends Terminator {
     if (farmResultNodes.length > 0) {
       // Hide farm result nodes.
       farmResultNodes.forEach(resultNode => resultNode.classList.add('cft-farm-result'))
-      hideElements(farmResultNodes, false) // fire and forget
+      this.hideResults(farmResultNodes, true) // fire and forget
+      // hideElements(farmResultNodes, false) // fire and forget
       // Add a hint which allows user to show these results temporarily onto the
       // page.
       this.addShowFarmResultsOnceHint(farmResultNodes.length)
@@ -77,27 +78,29 @@ export abstract class ListedTerminator extends Terminator {
         // Mark the result node as shown temporarily
         resultNode.classList.add('cft-farm-result-shown')
 
-        button.onclick = once(async (e) => {
-          e.preventDefault()
+        if (button) {
+          button.onclick = once(async (e) => {
+            e.preventDefault()
 
-          // Get other search items linking to same host and show them all.
-          const domain = this.getSourceDomain(resultNode)
-          const unblockedResultNodes = this.getResultNodes().filter(resultNode => this.getSourceDomain(resultNode) === domain)
-          // Since these results are unblocked, we should re-add the
-          // "Terminate!" hint to them, and set title to safe (default font).
-          unblockedResultNodes.forEach(unblockedResultNode => {
-            unblockedResultNode.classList.remove('cft-farm-result', 'cft-farm-result-shown')
-            this.addTerminateHint(unblockedResultNode)
+            // Get other search items linking to same host and show them all.
+            const domain = this.getSourceDomain(resultNode)
+            const unblockedResultNodes = this.getResultNodes().filter(resultNode => this.getSourceDomain(resultNode) === domain)
+            // Since these results are unblocked, we should re-add the
+            // "Terminate!" hint to them, and set title to safe (default font).
+            unblockedResultNodes.forEach(unblockedResultNode => {
+              unblockedResultNode.classList.remove('cft-farm-result', 'cft-farm-result-shown')
+              this.addTerminateHint(unblockedResultNode)
+            })
+
+            // Remove this host to blocking list
+            await db.removeHosts([domain])
           })
-
-          // Remove this host to blocking list
-          await db.removeHosts([domain])
-        })
+        }
       })
 
       // Then we show the blocked results and hide the show-once hint.
       await Promise.all([
-        showElements(blockedResultNodes, true),
+        this.showResults(blockedResultNodes),
         hideElements([document.getElementById('cft-temp-show') as HTMLElement], true)
       ])
     })
@@ -107,45 +110,49 @@ export abstract class ListedTerminator extends Terminator {
     const domain = this.getSourceDomain(resultNode)
     const hintNode = this.addHintNode(resultNode, _('terminateHint'))
 
-    hintNode.onclick = once(async (e) => {
-      e.preventDefault()
+    if (hintNode) {
+      hintNode.onclick = once(async (e) => {
+        e.preventDefault()
 
-      // Get other search items linking to same host and hide them all.
-      const resultNodes = this.getResultNodes()
-      const nodesToBeHidden = resultNodes.filter(e => this.getSourceDomain(e) === domain)
-      await hideElements(nodesToBeHidden, true) // wait for fade effect
+        // Get other search items linking to same host and hide them all.
+        const resultNodes = this.getResultNodes()
+        const resultNodesToBeHidden = resultNodes.filter(e => this.getSourceDomain(e) === domain)
+        await this.hideResults(resultNodesToBeHidden, false) // wait for fade effect
 
-      // THEN add undo hints to these nodes
-      nodesToBeHidden.forEach(e => this.addUndoHint(e))
+        // THEN add undo hints to these nodes
+        resultNodesToBeHidden.forEach(e => this.addUndoHint(e))
 
-      // Add this host to blocking list
-      await db.addHosts([domain])
-    })
+        // Add this host to blocking list
+        await db.addHosts([domain])
+      })
+    }
   }
 
   private addUndoHint(hiddenResultNode: HTMLElement): void {
     const domain = this.getSourceDomain(hiddenResultNode)
     const undoButton = this.addUndoHintNode(hiddenResultNode, _('undoHint'), _('terminatedMsg', domain))
 
-    undoButton.onclick = once(async (e) => {
-      e.preventDefault()
+    if (undoButton) {
+      undoButton.onclick = once(async (e) => {
+        e.preventDefault()
 
-      // Get all undo node for the domain and removes them
-      const undoNodes = document.querySelectorAll(`.cft-blocked-hint[cft-domain="${domain}"]`)
-      undoNodes.forEach(undoNode => undoNode.remove())
+        // Get all undo node for the domain and removes them
+        const undoNodes = document.querySelectorAll(`.cft-blocked-hint[cft-domain="${domain}"]`)
+        undoNodes.forEach(undoNode => undoNode.remove())
 
-      // Show hidden result nodes that are previous hidden.
-      let unblockedResultNodes = this.getResultNodes()
-      unblockedResultNodes = unblockedResultNodes.filter(resultNode => this.getSourceDomain(resultNode) === domain)
-      unblockedResultNodes = unblockedResultNodes.filter(isElementHidden)
-      await showElements(unblockedResultNodes, true)
+        // Show hidden result nodes that are previous hidden.
+        let unblockedResultNodes = this.getResultNodes()
+        unblockedResultNodes = unblockedResultNodes.filter(resultNode => this.getSourceDomain(resultNode) === domain)
+        // unblockedResultNodes = unblockedResultNodes.filter(isElementHidden) // what does this mean?
+        await this.showResults(unblockedResultNodes)
 
-      // Re-add 'Terminate!' button to these nodes
-      unblockedResultNodes.forEach(resultNode => this.addTerminateHint(resultNode))
+        // Re-add 'Terminate!' button to these nodes
+        unblockedResultNodes.forEach(resultNode => this.addTerminateHint(resultNode))
 
-      // Remove domain from database
-      await db.removeHosts([domain])
-    })
+        // Remove domain from database
+        await db.removeHosts([domain])
+      })
+    }
   }
 
   // Queries all result nodes.
@@ -164,17 +171,21 @@ export abstract class ListedTerminator extends Terminator {
   //   is set to "Terminate!" as same as the first one.
   //
   // Returns the button element.
-  protected abstract addHintNode(resultNode: HTMLElement, text: string): HTMLElement
+  protected abstract addHintNode(resultNode: HTMLElement, text: string): HTMLElement | null
 
   // Add an undo hint and button onto the page. Returns the button element. The
   // result node should be invisible but remains in the DOM tree. Do not add
   // click listener for the undo button. It is super class's responsibility.
-  protected abstract addUndoHintNode(hiddenResultNode: HTMLElement, buttonText: string, undoHintText: string): HTMLElement
+  protected abstract addUndoHintNode(hiddenResultNode: HTMLElement, buttonText: string, undoHintText: string): HTMLElement | null
 
   // Add a hint banner at the end of the webpage. Returns the button to be
   // click. Do not add listener to the button. It is super class's
   // responsibility.
   protected abstract addShowFarmResultsOnceNode(msgLeft: string, buttonText: string, msgRight: string): HTMLElement
+
+  protected abstract hideResults(elements: HTMLElement[], init: boolean): Promise<void>
+
+  protected abstract showResults(elements: HTMLElement[]): Promise<void>
 }
 
 // Unlisted terminator is used for webpage whose search results are not a
@@ -191,6 +202,7 @@ export abstract class UnlistedTerminator extends Terminator {
     // Since search results are dynamically loaded, an mutation observer is
     // needed.
     const ma = new MutationObserver((muts) => {
+      let addedFarmResults: HTMLElement[] = []
       for (const mut of muts) {
         for (let i = 0; i < mut.addedNodes.length; i++) {
           const addedNode = mut.addedNodes.item(i)
@@ -198,22 +210,25 @@ export abstract class UnlistedTerminator extends Terminator {
             addedNode.classList.add('cft-result')
             const domain = this.getSourceDomain(addedNode)
             if (farmList.has(domain)) {
+              addedFarmResults.push(addedNode)
               addedNode.classList.add('cft-farm-result')
               // hideElements([addedNode], false) // fire and forget
             }
           }
         }
       }
+      greyOutElements(addedFarmResults, true)
     })
     const resultContainer = this.getSearchResultWrapper()
     ma.observe(resultContainer, { childList: true })
 
-    // Hide current appeared farm results
+    // Hide currently appeared farm results
     const resultNodes = this.getCurrentSearchResults()
     resultNodes.forEach(e => e.classList.add('cft-result'))
 
     const farmResultNodes = resultNodes.filter(e => farmList.has(this.getSourceDomain(e)))
     farmResultNodes.forEach(e => e.classList.add('cft-farm-result'))
+    greyOutElements(farmResultNodes, true)
     // hideElements(farmResultNodes, false) // fire and forget
 
     // Add option to allow user stops the terminator
@@ -225,18 +240,28 @@ export abstract class UnlistedTerminator extends Terminator {
 
       const tempHintNode = document.getElementById('cft-temp-show') as HTMLElement
       const hiddenFarmImageResults = Array.from(resultContainer.querySelectorAll('.cft-farm-result')) as HTMLElement[]
-      hiddenFarmImageResults.forEach(e => e.classList.remove('cft-farm-result'))
+      hiddenFarmImageResults.forEach(e => e.classList.add('cft-farm-result-shown'))
+      greyInElements(hiddenFarmImageResults, true) // fire and forget
+
       await hideElements([tempHintNode], true)
     })
   }
 
-  abstract getSearchResultWrapper(): HTMLElement
+  protected abstract getSearchResultWrapper(): HTMLElement
 
-  abstract isSearchResult(e: HTMLElement): boolean
+  protected abstract isSearchResult(e: HTMLElement): boolean
 
-  abstract getCurrentSearchResults(): HTMLElement[]
+  protected abstract getCurrentSearchResults(): HTMLElement[]
 
-  abstract getSourceDomain(resultNode: HTMLElement): string
+  protected abstract getSourceDomain(resultNode: HTMLElement): string
 
-  abstract addCancelTerminatorHint(msgLeft: string, buttonText: string, msgRight: string): HTMLElement
+  protected abstract addCancelTerminatorHint(msgLeft: string, buttonText: string, msgRight: string): HTMLElement
+
+  protected hideElements(elements: HTMLElement[]): Promise<void> {
+    return greyOutElements(elements, true)
+  }
+
+  protected showElements(elements: HTMLElement[]): Promise<void> {
+    return greyInElements(elements, true)
+  }
 }
